Change Integer token values to unsigned longs.
There are exact negative float values (e.g. "-44.0") in the hirise
sample label, which suggests that integers are unsigned.

+++ src/token.h
@@ -66,7 +66,7 @@ typedef struct TknDate {
 typedef struct TknInteger {
     char type;
     char* str;
-    int intval;
+    unsigned long intval;
 } TknInteger;
 
 typedef struct TknRational {
